PAXOS LIVES!  HALLELUJAH TO THE GODS ALLMIGHTY

+++ sim_twitter.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
+if [ ! $# == 1 ] ; then
+  echo "Usage $0 [scripts/testname]"
+  exit
+fi
+
 rm -rf storage
 rm -f *.log
 rm -f *.replay
-./execute_eclipse.pl -s -n TwitterNode -f 1 -c scripts/CreateUser -l=blah
+./execute_eclipse.pl -s -n TwitterNode -f 1 -c $1 -l=blah
